requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
bcrypt==4.3.0
    # via -r requirements.in
celery[redis]
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
docker==7.1.0
    # via testcontainers
ecdsa==0.19.1
    # via python-jose
fastapi==0.117.1
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
hvac==2.3.0
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
packaging==25.0
    # via pytest
passlib==1.7.4
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via -r requirements.in
python-dotenv==1.1.1
    # via testcontainers
python-jose==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
redis==6.4.0
    # via -r requirements.in
requests==2.32.5
    # via
    #   docker
    #   hvac
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via -r requirements.in
starlette==0.48.0
    # via fastapi
testcontainers==4.13.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   testcontainers
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   docker
    #   requests
    #   testcontainers
wrapt==1.17.3
    # via testcontainers
python-dotenv ==1.1.1